Do not reuse the same image for all pr jobs
Because all jobs were using the same tag for the docker image, they
would end up writing over the image that was in use by another pr job.
This could cause code that didn't even exist in the PR to exist in the
test docker image.

Use an image named for the specific pr to avoid this contention.

Fixes #174